# api.yaml
openapi: 3.0.0
info:
  title: "Szabadságnyilvántartó API"
  version: 1.0.0
paths:
...
  "/uzenetek/{id}":
    {
      get:
        {
          summary: "Retrieve a list of messages for a specific user by ID",
          tags: [Messages],
          parameters:
            [
              {
                in: path,
                name: id,
                schema: { type: string },
                required: true,
                description: "The ID of the user to retrieve messages for",
              },
            ],
          responses:
            {
              "200":
                {
                  description: "Returns a status and a list of messages or an error message",
                  content:
                    {
                      application/json:
                        {
                          schema:
                            {
                              type: object,
                              properties:
                                {
                                  status:
                                    {
                                      type: string,
                                      description: "The status of the request",
                                    },
                                  messages:
                                    {
                                      type: array,
                                      items:
                                        {
                                          $ref: "#/components/schemas/Message",
                                        },
                                    },
                                  error:
                                    {
                                      type: string,
                                      description: "The error message, if any",
                                    },
                                },
                            },
                        },
                    },
                },
            },
        },
    }
  /uzenetek/create:
    {
      post:
        {
          summary: "Create a new message for a specific user based on the permissions of the submitting user",
          tags: [Messages],
          security: [{ ApiKeyAuth: [] }],
          requestBody:
            {
              required: true,
              content:
                {
                  application/json:
                    {
                      schema:
                        {
                          type: object,
                          properties:
                            {
                              userId: { type: string },
                              date: { type: string },
                              message: { type: string },
                            },
                        },
                    },
                },
            },
          responses:
            {
              "200":
                {
                  description: "Returns a status and a message or an error message",
                  content:
                    {
                      application/json:
                        {
                          schema:
                            {
                              type: object,
                              properties:
                                {
                                  status:
                                    {
                                      type: string,
                                      description: "The status of the request",
                                    },
                                  message:
                                    { $ref: "#/components/schemas/Message" },
                                  error:
                                    {
                                      type: string,
                                      description: "The error message, if any",
                                    },
                                },
                            },
                        },
                    },
                },
            },
        },
    }
components:
  securitySchemes:
    { ApiKeyAuth: { type: apiKey, in: header, name: submittingId } }
  schemas:
...
      Message:
        {
          type: object,
          properties:
            {
              userId: { type: string },
              date: { format: date },
              message: { type: string },
              sendingName: { type: string },
            },
        },
...